Fix many compiler errors

+++ src/Mewdeko/Modules/MultiGreets/SlashMultiGreets.cs
@@ -77,7 +77,7 @@ public class SlashMultiGreets : MewdekoSlashModuleBase<MultiGreetService>
 
     [SlashCommand("delete","Set how long it takes for a greet to delete"),  RequireContext(ContextType.Guild), SlashUserPerm(GuildPermission.Administrator),
      RequireBotPermission(GuildPermission.ManageMessages), CheckPermissions, BlacklistCheck]
-    public async Task MultiGreetDelete(int id, [Summary("Seconds", "After how long in seconds it should delete.")] ulong howlong)
+    public async Task MultiGreetDelete(int id, [Summary("Seconds", "After how long in seconds it should delete.")] int howlong)
     {
         var greet = Service.GetGreets(ctx.Guild.Id).ElementAt(id - 1);
         if (greet is null)
